streamlit_app/requirements.txt
streamlit
pandas~=2.2.3
numpy~=2.2.4
plotly~=6.0.1
plotly_express~=0.4.1
watchdog~=6.0.0
babel~=2.17.0
diskcache
streamlit-autorefresh
python-dateutil